pyproject.toml
[tool.poetry]
name = "pheval"
version = "0.5.0"
description = ""
authors = ["Yasemin Bridges <y.bridges@qmul.ac.uk>",
  "Julius Jacobsen <j.jacobsen@qmul.ac.uk>",
  "Nico Matentzoglu <nicolas.matentzoglu@gmail.com>",
  "Vinícius de Souza <souzadevinicius@gmail.com>"]
readme = "README.md"
packages = [{include = "pheval", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0.0"
jaydebeapi = ">=1.2.3"
tqdm = ">=4.64.1"
pandas = ">=1.5.1"
deprecation = ">=2.1.0"
click = ">=8.1.3"
class-resolver = ">=0.4.2"
phenopackets = "^2.0.2"
oaklib = ">=0.5.6"
google = "^3.0.0"
pyaml = "^21.10.1"
plotly = "^5.13.0"
seaborn = "^0.12.2"
matplotlib = "^3.7.0"
pyserde = "^0.9.8"
polars = "^1.23"
scikit-learn = "^1.4.0"
duckdb = "^1.0.0"
pyarrow = "^19.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
coverage = "^6.5.0"
pytest-workflow = "^2.0.1"

[tool.poetry.scripts]
pheval = "pheval.cli:pheval"
pheval-utils = "pheval.cli:pheval_utils"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^8.5.11"
mkdocstrings = "^0.24.0"
mkdocstrings-python = "^1.8.0"
pymdown-extensions = "^9.9"
mkdocs-include-dir-to-nav = "^1.2.0"
mkdocs-click = "^0.8.0"
griffe = "0.38.1"
[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[tool.black]
line-length = 100
target-version = ["py39", "py310"]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
include_trailing_comma = true
reverse_relative = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
